Исправление ошибки SDK-326

+++ LiveTex.SampleApp/View/RequestDialogPage.xaml.cs
@@ -31,6 +31,7 @@ namespace LiveTex.SampleApp
 			base.OnNavigatedFrom(e);
 
 			if(e.NavigationMode == NavigationMode.New
+				&& e.Uri.OriginalString.Contains("DialogPage.xaml")
 				&& e.IsNavigationInitiator)
 			{
 				NavigationService.RemoveBackEntry();
